+++ src/includes/classes/Core.php
@@ -1,0 +1,13 @@
+<?php
+declare (strict_types = 1);
+namespace WebSharks\Dicer;
+
+/**
+ * Core (back compat).
+ *
+ * @since 150601
+ */
+class Core extends Di
+{
+    // For back compat. only.
+}